Code refactoring
Deleted uneccessary localstorage sets and put todo component to new React Component
