+++ src/components/Sidebar.jsx
@@ -1,34 +1,8 @@
 import React from 'react';
 import './Sidebar.css'
-import bin from '../images/delete.svg'
+import ToDoElem from "./ToDoElem";
 
 const Sidebar = (props) => {
-
-    const ToDos = props.todos.map(todo=>{
-
-        //Function to set styles to ToDo object if it is chosen as current ToDo
-
-        function isCurrent(prop, id) {
-            if (prop.currentTodo) return (prop.currentTodo.id === id? 'current' :'')
-            else return ''
-        }
-
-        return <div className='todo--comp' key={todo.id}>
-            <div
-            className={`todo--object ${isCurrent(props, todo.id)} ${todo.stateTodo}`}
-            onClick={() => props.setCurrent(todo.id)}
-        >
-            {todo.title}
-
-            </div>
-            <img
-                src={bin}
-                className={`todo--delete ${isCurrent(props, todo.id)} ${todo.stateTodo}`}
-                onClick={() => props.delete(todo.id)}
-            />
-
-        </div>
-    })
 
     return (
         <div className='todos--holder'>
@@ -51,7 +25,11 @@ const Sidebar = (props) => {
                 onChange={props.findTodos}
             />
             <div className='todo--list'>
-            {ToDos}
+            <ToDoElem
+                todos={props.todos}
+                currentTodo={props.currentTodo}
+                setCurrent={props.setCurrent}
+            />
             </div>
         </div>
     );
